Merge develop to main

+++ core-service/src/main/java/com/stride/tracking/coreservice/utils/calculator/heartrate/HeartRateCalculatorResult.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public record HeartRateCalculatorResult(
         Double avgHeartRate,
-        Double maxHeartRate,
+        Integer maxHeartRate,
         List<HeartRateZoneValue> heartRateZones) {
 }
